fixed bug in delete booking

+++ controllers/bookings.controller.js
@@ -59,21 +59,27 @@ const create = (req, res, next) => {
 }
 
 const destroy = (req, res) => {
-  const query = 'DELETE FROM bookings WHERE id = ?'
   const params = [req.params.id]
-  const user_id = parseInt(params[0])
-  if (user_id === req.userId) {
-    db.run(query, params, (err, result) => {
-      if(err) {
-        res.status(400).json({"error": "A database error occurred"})
-        return;
+  db.get('SELECT user FROM bookings WHERE id = ?', params, (err, booking) => {
+    if (booking) {
+      const query = 'DELETE FROM bookings WHERE id = ?'
+      const booking_user_id = booking.user
+      const logged_user_id = req.userId
+      if (booking_user_id === logged_user_id) {
+        db.run(query, params, (err, result) => {
+          if(err) {
+            res.status(400).json({"error": "A database error occurred"})
+            return;
+          }
+          res.status(200).json({ "message": "deleted" })
+        });
+      } else {
+        res.status(403).json({ "message": "403 Unauthorized action" })
       }
-      res.status(200).json({ "message": "deleted" })
-    });
-  } else {
-    res.status(403).json({ "message": "403 Unauthorized action" })
-  }
-
+    } else {
+      res.status(404).json({ "message": "Booking not found" })
+    }
+  })
 }
 
 module.exports = {
